Fixed type of accountID member in GC_GameServer_AuthChallengeResponse_t
--HG--
extra : convert_revision : svn%3Af45044ad-9681-42ff-a337-a86ec09c9930%40304

+++ Open Steamworks/GameCoordinatorCommon.h
@@ -473,7 +473,7 @@ struct GC_GameServer_AuthChallengeResponse_t
 	enum { k_iMessage = k_EMsgGC_GameServer_AuthChallengeResponse };
 	uint16 id;
 	char garbage[16];
-	uint16 accountID;
+	uint32 accountID;
 	// Variable length data:
 	// char hash[];
 }
